requirements.txt
rembg>=2.0.50
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
